pyproject.toml
[project]
name = "habit-tracker-backend"
version = "0.1.0"
description = "FastAPI backend for the habit tracker PWA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.0",
  "pymysql>=1.1.0",
  "bcrypt>=4.0.0",
  "pyjwt>=2.8.0",
  "email-validator>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
